Fix: disallow passing multiple test files to PHPUnit via files()
Unfortunately, PHPUnit does not support passing a list of test files
to run via the command-line interface. However, the current
implementation of the PHPUnit task allowed passing an array of files
to its `files()` method, which it would then append to the command as
a string of comma-delimited files, which would cause the PHPUnit
executable to fail.

The PHPUnit executable supports either a path to a directory
containing multiple test files or a path to a single test file as a
command-line argument.

This fixes the issue by allowing only a single string to be passed via
`files()`.

Signed-off-by: Antoine Busque <antoine.busque@pillrcompany.com>