Add failGroup() method for Codeception run command (#622)
* Add failGroup() method for Codeception run command

This feature is major for parallel running. Without this, codeception
puts failed tasks to the same file, each thread rewrites file. With
this config we can send fails to different files and aggregate them
later manually.

* Fix failGroup() method for Codeception run command

This feature is major for parallel running. Without this, codeception
puts failed tasks to the same file, each thread rewrites file. With
this config we can send fails to different files and aggregate them
later manually.

* Add failGroup() method for Codeception run command

This feature is major for parallel running. Without this, codeception
puts failed tasks to the same file, each thread rewrites file. With
this config we can send fails to different files and aggregate them
later manually.

* Fix tests for multiple platforms

Sometimes command generates with ' surrounding arguments and sometimes
it is a ". Now used regexp to handle both styles