Fix the main Robo script entrypoint to work as a phar. (#811)
\Phar::running() is documented to return nothing in the stub. This method DOES
work when the phar is built with https://github.com/humbug/box, but it behaves
as documented when Robo builds its phar with the standard Phar APIs (which is
to say that it returns an empty string).

When running as a phar, __FILE__ will return whatever the current filename for
the phar is, so the documentation's suggestion to use __FILE__ in place of
\Phar::running() also does not work. However, the path that we compiled the
phar with, 'phar://robo.phar/...' always works.  We therefore use that first
if it exists.