[BUGFIX] Remove redundant check preventing file writes
Remove the check that a given filename is writetable,
since this requires the file to exist beforehand.

We dont expect the file to exist however. Therefore this
check should not be made or create a file if failing.

Since this check is included in file_put_contents anyway
we may remove the check safely.

Closes #959