[FEATURE] Replace lurker package
The watch task requires the package »henrikbjorn/lurker«.
The dependencies of this package dont allow using Symfony 4.
Last release of this package was 2016-03-16. So its abdandoned.

Use the drop-in replacement package totten/lurkerlite instead.
Installing this packsge will remove »henrikbjorn/lurker« automatically.

Since the package is a designated drop-in replacement all existing
methods in robo may stay the same.

Closes #459 #956 #973