[BUGFIX] Restore symfony/yaml dependency (#1010)
Robos configuration loader uses the symfony/yaml package
to parse YAML configuration files. This dependency was removed unintentionally
when package "grasmash/yaml-expander" was removed. The tests stayed green
because another development package still required the yaml parser.

Restore the dependency to avoid fatal errors in production.

Support version 4 and 5 of the package, as described in the robo branch overview.

Closes #1009