[FEATURE] Image Minify: Skip test on windows
Right now it is not possible to run the test on Windows
due to several issues in the imagemin downloader.

The task runs without any minification on windows.
So the following checks are superfluous.

Therefore it should be skipped for now on win.
The linux tests still work as expected.

Refs #1052 #1053