[BUGFIX] Image Minify: Skip files without minifier
Early return for all files which have no minifier set,
to avoid fatal error with ambigous error message
“Minifier method  cannot be found!”.
For example when webp images exist in the image directory.
The files are added to the error list right away,
to that the task may continue to minify other images and
show a result message (“Minified X of Y files”) at the end.